Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > TMPDIR is set to /data/local/tmp in external/bash/etc/bashrc.  So you could add "mkdir -p $TMPDIR" there.
> 
> I thought recovery tmp is /tmp and that's what should be used

Doesn't matter as long as ${TMPDIR:-/tmp} exists.

Sure recovery mounts /tmp as tmpfs and it's perfectly reasonable to use it.  That would require either not setting TMPDIR or setting it to /tmp in bashrc when in recovery mode. So maybe something like this:

-export TMPDIR=/data/local/tmp
+[-z "$(getprop init.svc.recovery)"] && export TMPDIR=/data/local/tmp

Another idea: symlink /data/local/tmp -> /tmp in recovery.  Though I'd prefer not to do that.

Patch-set: 1