Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > TMPDIR is set to /data/local/tmp in external/bash/etc/bashrc.  So you could add "mkdir -p $TMPDIR" there.
> > 
> > I thought recovery tmp is /tmp and that's what should be used
> 
> Doesn't matter as long as ${TMPDIR:-/tmp} exists.
> 
> Sure recovery mounts /tmp as tmpfs and it's perfectly reasonable to use it.  That would require either not setting TMPDIR or setting it to /tmp in bashrc when in recovery mode. So maybe something like this:
> 
> -export TMPDIR=/data/local/tmp
> +[-z "$(getprop init.svc.recovery)"] && export TMPDIR=/data/local/tmp
> 
> Another idea: symlink /data/local/tmp -> /tmp in recovery.  Though I'd prefer not to do that.

The thing is that we're using mksh in recovery too, so anything appended to bash won't change things.
The TMPDIR variable right now seems to be set by system/core and not via a ramdisk.
Yesterday I tried to move this to init.rc for both recovery and system, but that didn't seem to work at least for the former, the variable was just empty/unset

Patch-set: 1
Reviewer: Gerrit User 18106 <18106@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>